+++ laserchicken/feature_extractor/__init__.py
@@ -1,6 +1,4 @@
 """Feature extractor module."""
-import importlib
-import re
 
 import numpy as np
 import sys
@@ -8,47 +6,19 @@ import sys
 import time
 
 from laserchicken import keys, utils
-from .density_feature_extractor import PointDensityFeatureExtractor
-from .echo_ratio_feature_extractor import EchoRatioFeatureExtractor
-from .eigenvals_feature_extractor import EigenValueVectorizeFeatureExtractor
-from .entropy_z_feature_extractor import EntropyZFeatureExtractor
-from .percentile_z_feature_extractor import PercentileZFeatureExtractor
-from .pulse_penetration_feature_extractor import PulsePenetrationFeatureExtractor
-from .sigma_z_feature_extractor import SigmaZFeatureExtractor
-from .median_z_feature_extractor import MedianZFeatureExtractor
-from .range_z_feature_extractor import RangeZFeatureExtractor
-from .var_z_feature_extractor import VarianceZFeatureExtractor
-from .mean_std_coeff_z_feature_extractor import MeanStdCoeffZFeatureExtractor
-from .skew_z_feature_extractor import SkewZFeatureExtractor
-from .kurtosis_z_feature_extractor import KurtosisZFeatureExtractor
-from .skew_norm_z_feature_extractor import SkewNormZFeatureExtractor
-from .mean_std_coeff_norm_z_feature_extractor import MeanStdCoeffNormZFeatureExtractor
-from .var_norm_z_feature_extractor import VarianceNormZFeatureExtractor
-from .range_norm_z_feature_extractor import RangeNormZFeatureExtractor
-from .kurtosis_norm_z_feature_extractor import KurtosisNormZFeatureExtractor
-from .entropy_norm_z_feature_extractor import EntropyNormZFeatureExtractor
-from .median_norm_z_feature_extractor import MedianNormZFeatureExtractor
-from .percentile_norm_z_feature_extractor import PercentileNormZFeatureExtractor
-from .density_absolute_mean_z_feature_extractor import DensityAbsoluteMeanZFeatureExtractor
-from .density_absolute_mean_norm_z_feature_extractor import DensityAbsoluteMeanNormZFeatureExtractor
+from laserchicken.feature_extractor.feature_map import create_default_feature_map, _create_name_extractor_pairs
+from laserchicken.feature_extractor.base_feature_extractor import FeatureExtractor
+
+FEATURES = create_default_feature_map()
 
 
-def _create_feature_map(module_name=__name__):
-    """Construct a mapping from feature names to feature extractor classes."""
-    name_extractor_pairs = _find_name_extractor_pairs(module_name)
-    return {feature_name: extractor for feature_name, extractor in name_extractor_pairs}
+def list_feature_names():
+    return FEATURES#[feature_name for feature_name in FEATURES]
 
 
-def _find_name_extractor_pairs(module_name):
-    module = importlib.import_module(module_name)
-    name_extractor_pairs = [(feature_name, extractor)
-                            for name, extractor in vars(module).items() if
-                            re.match('^[A-Z][a-zA-Z0-9_]*FeatureExtractor$', name)
-                            for feature_name in extractor.provides()]
-    return name_extractor_pairs
-
-
-FEATURES = _create_feature_map()
+def register_new_feature_extractor(extractor: FeatureExtractor):
+    for name, extractor in _create_name_extractor_pairs([extractor]):
+        FEATURES[name] = extractor
 
 
 def compute_features(env_point_cloud, neighborhoods, target_idx_base, target_point_cloud, feature_names, volume,
@@ -91,7 +61,7 @@ def compute_features(env_point_cloud, neighborhoods, target_idx_base, target_poi
         if (target_idx_base == 0) and (not overwrite) and (feature_name in target_point_cloud[keys.point]):
             continue  # Skip feature calc if it is already there and we do not overwrite
 
-        extractor = FEATURES[feature_name]()
+        extractor = FEATURES[feature_name]
 
         if verbose:
             sys.stdout.write('Feature(s) "{}"'.format(extractor.provides()))
@@ -207,7 +177,7 @@ def _remove_duplicates(feature_list):
 def _make_extended_feature_list_helper(feature_names):
     feature_list = feature_names
     for feature_name in feature_names:
-        extractor = FEATURES[feature_name]()
+        extractor = FEATURES[feature_name]
         dependencies = extractor.requires()
         feature_list.extend(dependencies)
         feature_list.extend(_make_extended_feature_list_helper(dependencies))