+++ laserchicken/feature_extractor/percentile_z_feature_extractor.py
@@ -1,14 +1,15 @@
 import scipy.stats.stats as stats
 
-from laserchicken.feature_extractor.abc import AbstractFeatureExtractor
+from laserchicken.feature_extractor.base_feature_extractor import FeatureExtractor
 from laserchicken.keys import point
 
-PERCENTILES = range(10, 110, 10)
 
-
-class PercentileZFeatureExtractor(AbstractFeatureExtractor):
+class PercentileZFeatureExtractor(FeatureExtractor):
     """Height percentiles feature extractor class."""
     DATA_KEY = 'z'
+
+    def __init__(self, percentile=50):
+        self.percentile = percentile
 
     @classmethod
     def requires(cls):
@@ -22,8 +23,7 @@ class PercentileZFeatureExtractor(AbstractFeatureExtractor):
         """
         return []
 
-    @classmethod
-    def provides(cls):
+    def provides(self):
         """
         Get a list of names of the feature values.
 
@@ -33,7 +33,10 @@ class PercentileZFeatureExtractor(AbstractFeatureExtractor):
 
         :return: List of feature names
         """
-        return ['perc_{}_z'.format(i) for i in PERCENTILES]
+        return [self.generate_feature_name(self.percentile)]
+
+    def generate_feature_name(self, percentile):
+        return 'perc_{}_{}'.format(percentile, self.DATA_KEY)
 
     def extract(self, point_cloud, neighborhood, target_point_cloud, target_index, volume_description):
         """
@@ -47,7 +50,7 @@ class PercentileZFeatureExtractor(AbstractFeatureExtractor):
         :return: feature value
         """
         z = point_cloud[point][self.DATA_KEY]['data'][neighborhood]
-        return [stats.scoreatpercentile(z, p) for p in PERCENTILES]
+        return stats.scoreatpercentile(z, self.percentile)
 
     def get_params(self):
         """