+++ laserchicken/feature_extractor/test_all_features.py
@@ -1,21 +1,17 @@
-import importlib
 import json
 import os
 
 import pytest
-import numpy as np
+
 from laserchicken import compute_neighbors
-from laserchicken import keys
-from laserchicken import read_las
 from laserchicken import read_ply
+from laserchicken.feature_extractor import *
 from laserchicken.feature_extractor.pulse_penetration_feature_extractor import GROUND_TAGS
 from laserchicken.keys import point, normalized_height
 from laserchicken.utils import copy_point_cloud
-from laserchicken.volume_specification import InfiniteCylinder
-from laserchicken.feature_extractor import *
-
-from . import _create_feature_map, _find_name_extractor_pairs
+from laserchicken.volume_specification import InfiniteCylinder, Cell
 from . import compute_features
+from .feature_map import create_default_feature_map, _create_name_extractor_pairs
 
 np.random.seed(1234)
 
@@ -36,11 +32,12 @@ _10_NEIGHBORHOODS_IN_260807 = next(
 _260807_NEIGHBORHOODS_IN_10 = next(
     compute_neighbors.compute_neighborhoods(_PC_10, _PC_260807, _CYLINDER, sample_size=500))
 
-feature_names = [name for name in _create_feature_map()]
+features_by_name = create_default_feature_map()
+feature_names = [name for name in features_by_name]
 
 
 def test_no_duplicate_feature_registrations():
-    pairs = _find_name_extractor_pairs(__name__)
+    pairs = _create_name_extractor_pairs()
     for name, _ in pairs:
         matches = [extractor for extractor_name, extractor in pairs if extractor_name is name]
         np.testing.assert_equal(len(matches), 1,
@@ -105,20 +102,20 @@ def test_zeroPoints_consistentOutput(feature):
 
 @pytest.mark.parametrize("feature", feature_names)
 def test_zeroNeighbors_consistentOutput(feature):
-    _test_consistent_output_with_n_neighbors(feature, 0)
+    _assert_consistent_output_with_n_neighbors(feature, 0)
 
 
 @pytest.mark.parametrize("feature", feature_names)
 def test_oneNeighbor_consistentOutput(feature):
-    _test_consistent_output_with_n_neighbors(feature, 1)
+    _assert_consistent_output_with_n_neighbors(feature, 1)
 
 
 @pytest.mark.parametrize("feature", feature_names)
 def test_twoNeighbors_consistentOutput(feature):
-    _test_consistent_output_with_n_neighbors(feature, 2)
+    _assert_consistent_output_with_n_neighbors(feature, 2)
 
 
-def _test_consistent_output_with_n_neighbors(feature, n_neighbors):
+def _assert_consistent_output_with_n_neighbors(feature, n_neighbors):
     n_points = 10
     pc = _create_point_cloud(n=n_points)
     compute_features(pc, [range(n_neighbors) for _ in range(n_points)], 0, pc, [feature], volume=_CYLINDER)